bin/azure-dns-updater/requirements.txt
azure-identity>=1.5.0
azure-mgmt-dns>=2.0.0